requirements.txt
aiofiles
fastapi
pydantic-settings
twilio
uvicorn
websockets